requirements.txt
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
idna==2.10
lxml==4.5.2
MegafonAPI==0.0.41
multidict==5.1.0
pika==1.1.0
pika-stubs==0.1.3
pyquery==1.4.1
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.6.3
